% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_AULCSF.R
\name{sm_AULCSF}
\alias{sm_AULCSF}
\title{Calculation of the Area under a Curve (area under the log CSF, i.e., AULCSF)}
\usage{
sm_AULCSF(
  x,
  y,
  logXY = TRUE,
  n = 500,
  param0 = log10(c(100, 1, 2, 0.5)),
  param_upLimit = log10(c(2000, 20, 9, 2)),
  param_lowLimit = log10(c(2, 0.2, 1, 0.02))
)
}
\arguments{
\item{x}{This is the scalar spacing of the coordinate, such as the spatial frequency.
It is recommended that spatial frequency in linear units is provided.}

\item{y}{Numerical data of contrast sensitivity. The length of x and y must be equal.
It is recommended that contrast sensitivity in linear units is provided.}

\item{logXY}{When this argument is set to TRUE, the function will convert the x and y values
into log10 units, and then perform the integration. If it is set to FALSE,
the fuction will keep the data values as they are.}

\item{n}{Number of predicted data that are generated from the given model.}

\item{param0}{A vector that contains initial four parameters in log10 units in this order:
\enumerate{
\item peak gain, 2) peak spatial frequency, 3) bandwidth, and
\item truncation value.
}}

\item{param_upLimit}{A vector that contains the upper limit of the four parameters (in log10 units)
in this order: 1) peak gain, 2) peak spatial frequency, 3) bandwidth, and
4) truncation value.}

\item{param_lowLimit}{A vector that contains the lower limit of the four parameters (in log10 units)
in this order: 1) peak gain, 2) peak spatial frequency, 3) bandwidth, and
4) truncation value.}
}
\description{
Calculation of the Area under a Curve (area under the log CSF, i.e., AULCSF)
}
\examples{
\dontrun{

x <- c(0.25,0.35,0.48,0.68,0.94,1.31,1.83,2.54) # spatial frequency
y <- c(141,172,190,187,164,129,90.3,57.1) # averaged contrast sensitivity
sm_AULCSF(x,y) # log10 x and log 10 y are used to calculate AULCSF
sm_AULCSF(x,y) #  linear x and y are used to calculate AULCSF

}

}
