% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_AULCSF_list.R
\name{sm_AULCSF_list}
\alias{sm_AULCSF_list}
\title{Calculating AULCSF (area under the log CSF)
across multiple conditions, groups and subjects}
\usage{
sm_AULCSF_list(
  data,
  subjects,
  groups,
  conditions,
  x,
  values,
  logXY = TRUE,
  n = 500,
  param0 = log10(c(100, 1, 2, 0.5)),
  param_upLimit = log10(c(2000, 20, 9, 2)),
  param_lowLimit = log10(c(2, 0.2, 1, 0.02))
)
}
\arguments{
\item{data}{Name of the variable that stores the data frame that contains
the columns with the specified column names. It is recommended that the
data frame contains data of contrast sensitivity and spatial frequency
in linear units.}

\item{subjects}{The name of the column of the data frame that contains subjects.
It must be strings.}

\item{groups}{The name of the column of the data frame that contains each group.
It must be strings.}

\item{conditions}{The name of the column of the data frame that contains each condition.
It must be strings.}

\item{x}{The name of the column of the data frame that contains the
x-axis points/x coordinates (i.e., spatial frequency) from
which the AULCSF can be calculated. It must be strings.
The column must not have characters.}

\item{values}{The name of the column of the data frame that contains the
actual data (i.e., contrast sensitivity), which are the y-axis
points from which the AULCSF can be calculated. It must be strings.}

\item{logXY}{When this argument is set to TRUE, the function will convert the x and y values
into log10 units, and then perform the integration. If it is set to FALSE,
the function will keep the data values as they are.}

\item{n}{Number of predicted data that are generated from the given model.}

\item{param0}{A vector that contains initial four parameters in log10 units in this order:
\enumerate{
\item peak gain, 2) peak spatial frequency, 3) bandwidth, and
\item truncation value.
}}

\item{param_upLimit}{A vector that contains the upper limit of the four parameters (in log10 units)
in this order: 1) peak gain, 2) peak spatial frequency, 3) bandwidth, and
4) truncation value.}

\item{param_lowLimit}{A vector that contains the lower limit of the four parameters (in log10 units)
in this order: 1) peak gain, 2) peak spatial frequency, 3) bandwidth, and
4) truncation value.}
}
\value{

}
\description{
@description
}
\details{
This function returns a data frame containing AULCSFs from a data frame
that contains the original raw data. One of the two arguments 'groups'
and 'conditions' must be filled. The function will throw an error
if both arguments are empty. If 'logXY' is set TRUE, it will convert the
x and y data into log10 units, and then perform the AULCSF calculation.
}
