% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_cutOff.R
\name{sm_cutOff}
\alias{sm_cutOff}
\title{Calculating log cutoff spatial frequency}
\usage{
sm_cutOff(logGain, logCenter, octaveWidth)
}
\arguments{
\item{logGain}{This refers to peak gain (i.e., peak sensitivity) of the contrast sensitivity function.
It should be provided in log10 units.}

\item{logCenter}{This refers to peak spatial frequency (i.e., center of the contrast sensitivity function).
It should be provided in log10 units.}

\item{octaveWidth}{This refers to the bandwidth of the contrast sensitivity function. It should be provided in log10
units.}
}
\description{
This is used to analyze whether sensitivity at higher spatial frequency
is improved. The higher the cutoff SF, the better the sensitivity at higher
spatial frequency. This value, by definition, is the spatial frequency where
log-sensitivity = 0.
}
