% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_fitCSF.R
\name{sm_fitCSF}
\alias{sm_fitCSF}
\title{Calculating the sum of squares with predicted values and the raw data of contrast
sensitivity}
\usage{
sm_fitCSF(param, x, y)
}
\arguments{
\item{param}{This is a vector that should contain four values in this order:
\enumerate{
\item peak gain (log10 units), 2) peak spatial frequency (log10 units),
\item bandwidth (log10 units), and 4) truncation value (log10 units).
}}

\item{x}{A vector of spatial frequencies from which the contrast sensitivity data
have been collected. This should be the linear values of the tested spatial frequencies, not log.}

\item{y}{Linear units of contrast sensitivity data, not log.}
}
\value{

}
\description{
Using the four parameters from the Quick-CSF method, this function
compares the differences between the raw data (y) and the predicted data. The
predicted data are calculated from the four parameters at the given
spatial frequencies (x). This difference is computed as the sum of squares.
If the sum of squares is low, then the predicted values from the given model
are highly representative of the actual raw data.
}
