% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_np_boot.R
\name{sm_np_boot}
\alias{sm_np_boot}
\title{Non-parametric bootstrapping with replacement}
\usage{
sm_np_boot(param_list, n = 51)
}
\arguments{
\item{param_list}{This is the result from sm_param_list(), which is a list that contains
qCSF parameters of all observers, experimental conditions and groups, as well
as the tested spatial frequencies.}

\item{n}{Number of bootstraps (i.e., simulated observers). The default is set to 51.}
}
\description{
This function randomly samples qCSF parameters from observers, then generates
fitted values as a function of spatial frequency using the sampled qCSF parameters.
It returns a matrix, which can then be directly used for functions that perform
factor analysis.
}
\examples{
\dontrun{

df <- read_csv('https://www.smin95.com/data_ACh.csv') \%>\%
group_by(Subject,SpatialFreq) \%>\%
 summarise(Sensitivity = mean(Sensitivity),
           Repetition = 'avg')

                     conditions = 'Repetition',
                     x = 'SpatialFreq',
                     values = 'Sensitivity',
                     data = df)

# samples 1000 times with replacement
sm_np_boot(res, 1000)
}

}
