% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_params_list.R
\name{sm_params_list}
\alias{sm_params_list}
\title{Calculation the four + one parameters of the contrast sensitivity function
across multiple groups, conditions and subjects.}
\usage{
sm_params_list(
  data,
  subjects,
  groups,
  conditions,
  x,
  values,
  param0 = log10(c(100, 1, 2, 0.5)),
  param_upLimit = log10(c(2000, 20, 9, 2)),
  param_lowLimit = log10(c(2, 0.2, 1, 0.02)),
  setLinSF = 1
)
}
\arguments{
\item{data}{Name of the variable that stores the data frame that contains
the columns with the specified column names. It is recommended that the
data frame contains data of contrast sensitivity and spatial frequency
in linear units.}

\item{subjects}{The name of the column of the data frame that contains subjects.
It must be strings.}

\item{groups}{The name of the column of the data frame that contains each group.
It must be strings.}

\item{conditions}{The name of the column of the data frame that contains each condition.
It must be strings.}

\item{x}{The name of the column of the data frame that contains the
x-axis points/x coordinates (i.e., spatial frequency) from
which the four parameters can be calculated. It must be strings.
The column must not have characters.}

\item{values}{The name of the column of the data frame that contains the
actual data (i.e., contrast sensitivity), which are the y-axis
points from which the four parameters can be calculated.
It must be strings.}

\item{param0}{A vector that contains initial four parameters in log10 units in this order:
\enumerate{
\item peak gain, 2) peak spatial frequency, 3) bandwidth, and 4) truncation value.
}}

\item{param_upLimit}{A vector that contains the upper limit of the four parameters (in log10 units)
in this order: 1) peak gain, 2) peak spatial frequency, 3) bandwidth, and
4) truncation value.}

\item{param_lowLimit}{A vector that contains the lower limit of the four parameters (in log10 units)
in this order: 1) peak gain, 2) peak spatial frequency, 3) bandwidth, and
4) truncation value.}

\item{setLinSF}{This refers to the sensitivity level that the user wishes to measure the cut-off spatial frequency.
The default is set to 1. This means that the cut-off SF is measured when the linear
sensitivity equals to 1.}
}
\description{
This function returns a data frame containing the four + one parameters of the
contrast sensitivity function based on the raw data in the given data frame.
One of the two arguments 'groups' and 'conditions' must be filled. The function will throw an error
if both arguments are empty.
}
\examples{
\dontrun{
set.seed(1)
x <- c(0.25,0.35,0.48,0.68,0.94,1.31,1.83,2.54) # spatial frequency
y <- c(141,172,190,187,164,129,90.3,57.1) # averaged contrast sensitivity
se <- c(9.6,11,11.1,9.9,7.9,6.1,4.8,3.8) # standard error
gr <- c(rep(1,length(x)),rep(2,length(x)))
subj <- rep(paste0('S',1:2),each=8) # subject number
df <- data.frame(subj=subj,x=x,y=y, se=se, gr=gr)
df[1:length(x)+1,-c(1,2,5)] <- df[1:length(x)+1,-c(1,2,5)]+round(rnorm(10),1)
sm_params_list(data=df, subjects = 'subj',
              groups='gr',x='x', values='y')
}

}
