% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_plot_boot.R
\name{sm_plot_boot}
\alias{sm_plot_boot}
\title{Plotting eigenvalues and their 95\% confidnece intervals from non-parametric
bootstrap}
\usage{
sm_plot_boot(
  boot_res,
  point_size = 2.8,
  dodge_width = 0.2,
  line_width = 0.8,
  colors = c("#414242", "#797f85"),
  shapes = c(16, 17),
  labels = c("Data", "Random"),
  legends_loc = c(0.55, 0.7)
)
}
\arguments{
\item{boot_res}{The returned data frame from \code{sm_np_boot()}.}

\item{point_size}{Size of the points.}

\item{dodge_width}{Width of the positional difference between the two plots (data and random.
)}

\item{line_width}{The width of the lines.}

\item{colors}{The colors of the plots.}

\item{shapes}{The shapes of the plots.}

\item{labels}{The labels for the legend.}

\item{legends_loc}{Location of the legend (x and y).}
}
\description{
This function plots eigenvalues and their 95\% confidence interval by using
the data frame that is returned from \code{sm_np_boot()}. However, the author suggests that
the user plot the eigenvalues rather than to use \code{sm_plot_boot()} for more
customization. This function is for demonstration purposes
}
\examples{
\dontrun{

df <- read_csv('https://www.smin95.com/data_ACh.csv') \%>\%
group_by(Subject,SpatialFreq) \%>\%
 summarise(Sensitivity = mean(Sensitivity),
           Repetition = 'avg')

                     conditions = 'Repetition',
                     x = 'SpatialFreq',
                     values = 'Sensitivity',
                     data = df)

boot_res <- sm_np_boot(res, n=51)
sm_plot_boot(boot_res)
}

}
