% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_r2.R
\name{sm_r2}
\alias{sm_r2}
\title{Calculating R-squared of the contrast sensitivity data by comparing them
to the predicted values based on the given model}
\usage{
sm_r2(
  x,
  y,
  param0 = log10(c(100, 1, 2, 0.5)),
  param_upLimit = log10(c(2000, 20, 9, 2)),
  param_lowLimit = log10(c(2, 0.2, 1, 0.02))
)
}
\arguments{
\item{x}{A vector of spatial frequencies from which the contrast sensitivity data
have been collected. This should be the linear values of the tested spatial frequencies, not log.}

\item{y}{Linear units of contrast sensitivity data, not log.}

\item{param0}{A vector that contains initial four parameters in log10 units in this order:
\enumerate{
\item peak gain, 2) peak spatial frequency, 3) bandwidth, and
\item truncation value.
}}

\item{param_upLimit}{A vector that contains the upper limit of the four parameters (in log10 units)
in this order: 1) peak gain, 2) peak spatial frequency, 3) bandwidth, and
4) truncation value.}

\item{param_lowLimit}{A vector that contains the lower limit of the four parameters (in log10 units)
in this order: 1) peak gain, 2) peak spatial frequency, 3) bandwidth, and
4) truncation value.}
}
\description{
This compares the log data of the contrast sensitivity function (x- and y-axes both log)
to the predicted values based on the contrast sensitivity function model. This fuction
enables the users to assess whether the given model is appropriate to plot the
contrast sensitivity function with their data.
}
\examples{
\dontrun{

x <- c(0.6,1,2,4,8,16,24,36) # spatial frequency
y <- c(30, 50, 35, 25, 6, 4, 2, 1) # contrast sensitivity

sm_r2(x,y)
}

}
