% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_r2_list.R
\name{sm_R2_list}
\alias{sm_R2_list}
\title{Calculating R-squared of the contrast sensitivity function fit
across multiple conditions, groups and subjects}
\usage{
sm_R2_list(data, subjects, groups, conditions, x, values)
}
\arguments{
\item{data}{Name of the variable that stores the data frame that contains
the columns with the specified column names. It is recommended that the
data frame contains data of contrast sensitivity and spatial frequency
in linear units.}

\item{subjects}{The name of the column of the data frame that contains subjects.
It must be strings.}

\item{groups}{The name of the column of the data frame that contains each group.
It must be strings.}

\item{conditions}{The name of the column of the data frame that contains each condition.
It must be strings.}

\item{x}{The name of the column of the data frame that contains the
x-axis points/x coordinates (i.e., spatial frequency) from
which the AULCSF can be calculated. It must be strings.
The column must not have characters.}

\item{values}{The name of the column of the data frame that contains the
actual data (i.e., contrast sensitivity), which are the y-axis
points from which the AULCSF can be calculated. It must be strings.}
}
\description{
This function returns a data frame containing r2s from a data frame
that contains the original raw data. One of the two arguments 'groups'
and 'conditions' must be filled.
}
\examples{
\dontrun{

set.seed(1)
x <- c(0.25,0.35,0.48,0.68,0.94,1.31,1.83,2.54) # spatial frequency
y <- c(141,172,190,187,164,129,90.3,57.1) # averaged contrast sensitivity
se <- c(9.6,11,11.1,9.9,7.9,6.1,4.8,3.8) # standard error
gr <- c(rep(1,length(x)),rep(2,length(x)))
subj <- rep(paste0('S',1:2),each=8) # subject number
df <- data.frame(subj=subj,x=x,y=y, se=se, gr=gr)
df[1:length(x)+1,-c(1,2,5)] <- df[1:length(x)+1,-c(1,2,5)]+round(rnorm(10),1)
df$gr <- factor(df$gr)
sm_r2_list(data=df, subjects = 'subj', groups ='gr',x='x', values='y')
}
}
