% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_trapz.R
\name{sm_trapz}
\alias{sm_trapz}
\title{Calculation of the Area under a Curve (Trapezoidal
numerical integration)}
\usage{
sm_trapz(x, y, logXY = TRUE)
}
\arguments{
\item{x}{This is the scalar spacing of the coordinate, such as the spatial frequency.
It is recommended that spatial frequency in linear units is provided.}

\item{y}{Numerical data of contrast sensitivity. The length of x and y must be equal.
It is recommended that contrast sensitivity in linear units is provided.}

\item{logXY}{When this argument is set to TRUE, the function will convert the x and y values
into log10 units, and then perform the integration. If it is set to FALSE,
the fuction will keep the data values as they are.}
}
\value{

}
\description{
Calculation of the Area under a Curve (Trapezoidal
numerical integration)
}
\examples{

}
