% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_trapz_list.R
\name{sm_trapz_list}
\alias{sm_trapz_list}
\title{Calculating Area under Curve (trapezoidal method)
across multiple conditions, groups and subjects}
\usage{
sm_trapz_list(data, subjects, groups, conditions, x, values, logXY = TRUE)
}
\arguments{
\item{data}{Name of the variable that stores the data frame that contains
the columns with the specified column names. It is recommended that the
data frame contains data of contrast sensitivity and spatial frequency
in linear units.}

\item{subjects}{The name of the column of the data frame that contains subjects.
It must be strings.}

\item{groups}{The name of the column of the data frame that contains each group.
It must be strings.}

\item{conditions}{The name of the column of the data frame that contains each condition.
It must be strings.}

\item{x}{The name of the column of the data frame that contains the
x-axis points/x coordinates (i.e., spatial frequency) from
which the AUC can be calculated. It must be strings.
The column must not have characters.}

\item{values}{The name of the column of the data frame that contains the
actual data (i.e., contrast sensitivity), which are the y-axis
points from which the AUC can be calculated. It must be strings.}

\item{logXY}{When this argument is set to TRUE, the function will convert the x and y values
into log10 units, and then perform the integration. If it is set to FALSE,
the function will keep the data values as they are.}
}
\value{

}
\description{
This function returns a data frame containing AUCs from a data frame
that contains the original raw data. One of the two arguments 'groups'
and 'conditions' must be filled. The function will throw an error
if both arguments are empty. If 'logXY' is set TRUE, it will convert the
x and y data into log10 units, and then perform the trapezoidal integration.
}
