% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_CSF.R
\name{sm_CSF}
\alias{sm_CSF}
\title{Contrast sensitivity function}
\usage{
sm_CSF(..., logXY = TRUE)
}
\arguments{
\item{...}{Other arguments that are passed to control for the appearance of the
contrast sensitivity function, such as
\code{color = sm_color('blue')}.}

\item{logXY}{When this argument is set to TRUE (default), the function will convert the x and y values
into log10 units, and then plot the function in log10 scales. If it is set to FALSE,
the function will keep the data values as they are in the linear scale.}
}
\description{
It plots the contrast sensitivity function.

The function's default converts the x and y coordinates into log scales.

If logXY = TRUE, the plotted function will be in log xy scales.
}
\examples{
\dontrun{
x <- c(0.6,1,2,4,8,16,24,36) # spatial frequency
y <- c(30, 50, 35, 25, 6, 4, 2, 1) # contrast sensitivity
df <- data.frame(x=x,y=y)

# linear plot of CSF
ggplot(data=df,aes(x,y)) +
geom_point() +
sm_CSF()
}
}
