% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_cutOff.R
\name{sm_cutOff}
\alias{sm_cutOff}
\title{Calculating log cutoff spatial frequency}
\usage{
sm_cutOff(logGain, logCenter, octaveWidth, sensCutOff = 1)
}
\arguments{
\item{logGain}{This refers to peak gain (i.e., peak sensitivity) of the contrast sensitivity function.
It should be provided in log10 units.}

\item{logCenter}{This refers to peak spatial frequency (i.e., center of the contrast sensitivity function).
It should be provided in log10 units.}

\item{octaveWidth}{This refers to the bandwidth of the contrast sensitivity function. It should be provided in log10
units.}

\item{sensCutOff}{This refers to the sensitivity level that the user wishes to measure the cut-off spatial frequency.
The default is set to 1. This means that the cut-off SF is measured when the linear
sensitivity equals to 1.}
}
\description{
This is used to analyze whether sensitivity at higher spatial frequency
is improved. The higher the cutoff SF, the better the sensitivity at higher
spatial frequency. This value, by definition, is the spatial frequency where
log-sensitivity = 0.
}
