% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_facet_header.R
\name{sm_facet_header}
\alias{sm_facet_header}
\title{Customising the aesthetics of the facet header.}
\usage{
sm_facet_header(font_size = 10, head_fill = "white", font_weight = "bold")
}
\arguments{
\item{font_size}{Font size of the facet header. Number referring to the font size
should be provided.}

\item{head_fill}{The color that will fill the head of each facet. The hex code of a color,
or a character string of the color, should be provided here.}

\item{font_weight}{The font weight, such as 'bold', 'plain', 'italic', 'bold.italic' can
be specified here.}
}
\value{

}
\description{
Customising the aesthetics of the facet header.
}
