% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_facet_label.R
\name{sm_facet_label}
\alias{sm_facet_label}
\title{Facet header labeling}
\usage{
sm_facet_label(str1, sep_by = " ", df_column, sep_by2 = NULL, str2 = NULL)
}
\arguments{
\item{str1}{Character string that precedes the factor level as specified by the
column of the data frame \code{df_column}.}

\item{sep_by}{Character string that separates \code{str1} from the factor level as specified
by the column of the data frame \code{df_column}.}

\item{df_column}{Column of the data frame that provides the factor levels.}

\item{sep_by2}{Character string that separates \code{str2} from the factor level as specified
by the column of the data frame \code{df_column}.}

\item{str2}{Character string that comes after the factor level as specified by the
column of the data frame \code{df_column}.}
}
\value{

}
\description{
Facet header labeling
}
