% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_np_boot.R
\name{sm_np_boot}
\alias{sm_np_boot}
\title{Non-parametric bootstrap factor analysis for contrast sensitivity data}
\usage{
sm_np_boot(param_list, n = 50, nSim = 1000, ci_range = 0.95, fa = TRUE)
}
\arguments{
\item{param_list}{This is the result from sm_param_list(), which is a list that contains
qCSF parameters of all observers, experimental conditions and groups, as well
as the tested spatial frequencies.}

\item{n}{Number of sample for bootstrap. The default is set to 50.}

\item{nSim}{Number of non-parametric bootstrap simulations (number of samplings)}

\item{ci_range}{Range of the confidence interval. The default is set to 0.95.}

\item{fa}{If set to TRUE (default), it computes eigenvalues from the factor model,
which accounts for the random error. If FALSE, it computes from the principal component.}
}
\description{
This function randomly samples qCSF parameters from observers, then generates
fitted values as a function of spatial frequency using the sampled qCSF parameters.
Then using the fitted values, it performs parallel analysis to generate eigenvalues
both from the data and random matrices. For more information about parallel analysis
please check out \code{fa.parallel} from \code{psych} package.
}
\details{
It returns a data frame, which can then be directly used to plot the eigenvalues
using \code{sm_plot_boot}, which is an wrapper around ggplot2.

Essentially, it computes the eigenvalues from factor analysis/principal component for many
iterations, and the returns the mean and the 95\% confidence intervals of the
eigenvalues. Eigenvalues represent the proportion of the total variance that can be
described by a specific factor/principal component. The higher the eigenvalue of the factor/
component, the more important it is in the data.

If the eigenvalues' 95\% confidence interval from the bootstrap does not overlap with
the eigenvalues from the random matrix (i.e., significantly higher than the eigenvalues from the
random matrix), then it means that the factors are significant
and should be included in the model/dimension reduction.
}
\examples{
\dontrun{

df <- read_csv('https://www.smin95.com/data_ACh.csv') \%>\%
group_by(Subject,SpatialFreq) \%>\%
 summarise(Sensitivity = mean(Sensitivity),
           Repetition = 'avg')

                     conditions = 'Repetition',
                     x = 'SpatialFreq',
                     values = 'Sensitivity',
                     data = df)

sm_np_boot(res, n=51)
}

}
