% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_trapz.R
\name{sm_trapz}
\alias{sm_trapz}
\title{Calculation of the Area under a Curve (Trapezoidal
numerical integration)}
\usage{
sm_trapz(x, y, logXY = TRUE)
}
\arguments{
\item{x}{This is the scalar spacing of the coordinate, such as the spatial frequency.
It is recommended that spatial frequency in linear units is provided.}

\item{y}{Numerical data of contrast sensitivity. The length of x and y must be equal.
It is recommended that contrast sensitivity in linear units is provided.}

\item{logXY}{When this argument is set to TRUE (default), the function will convert the x and y values
into log10 units, and then perform the integration. If it is set to FALSE,
the fuction will keep the data values as they are.}
}
\description{
The default setting of the function converts the x and y data into
log units. If this is not wanted, set \code{logXY = FALSE}.
}
\examples{
\dontrun{

# trapezoidal area from linear spatial frequencies
# and contrast sensitivities
x <- c(0.6,1,2,4,8,16,24,36) # spatial frequency
y <- c(30, 50, 35, 25, 6, 4, 2, 1) # contrast sensitivity
sm_trapz(x,y,logXY=F)

# trapezoidal area from log10 spatial frequencies and
# log10 contrast sensitivities
sm_trapz(x,y) #logXY = TRUE is the default
}


}
